prepared parsing for map handling. clarified variables and functions names.
