+++ parsing/elements.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   init_data.c                                        :+:      :+:    :+:   */
+/*   elements.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mbecker <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/29 15:28:04 by mbecker           #+#    #+#             */
-/*   Updated: 2024/08/30 17:07:20 by mbecker          ###   ########.fr       */
+/*   Updated: 2024/08/30 17:24:06 by mbecker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,6 +36,73 @@ int	rgb_to_int(char *rgb_str)
 	return (rgb_int);
 }
 
+/**
+ * @brief Check the validity of a texture path.
+ * 
+ * @param elem_name The name of the element being checked.
+ * @param color The color string to be checked.
+ * @return 0 if the path is valid, otherwise 1.
+ */
+int	check_path(char *elem_name, char *path)
+{
+	int		fd;
+	char	*ptr;
+
+	if (!path)
+		return (error(elem_name, "no path found"));
+	ptr = ft_strrstr(path, ".xpm");
+	if (!ptr || ft_strlen(ptr) != 4)
+		return (error(path, "invalid texture file"));
+	fd = open(path, O_RDONLY);
+	if (fd == -1)
+		return (error(path, strerror(errno)));
+	close(fd);
+	return (0);
+}
+
+/**
+ * @brief Check the validity of a color string.
+ * 
+ * @param elem_name The name of the element being checked.
+ * @param color The color string to be checked.
+ * @return 0 if the color is valid, otherwise 1.
+ */
+int	check_color(char *elem_name, char *color)
+{
+	char	**rgb;
+	int		i;
+
+	if (!color)
+		return (error(elem_name, "no color found"));
+	if (ft_strlen(color) > 11)
+		return (error(elem_name, "invalid color format (string too long)"));
+	rgb = ft_split(color, ',');
+	if (!rgb || ft_tablen(rgb) != 3)
+	{
+		freetab(rgb, TRUE);
+		return (error(elem_name, "invalid color format (3 values required)"));
+	}
+	i = -1;
+	while (++i < 3)
+	{
+		if (!ft_isnum(rgb[i]) || ft_atoi(rgb[i]) < 0 || ft_atoi(rgb[i]) > 255)
+		{
+			freetab(rgb, TRUE);
+			error(elem_name, "invalid color format (Out of range or NaN)");
+			return (1);
+		}
+	}
+	freetab(rgb, TRUE);
+	return (0);
+}
+
+/**
+ * @brief Checks the validity of an element and updates the data structure 
+ * accordingly.
+ * @param val The element value in the format "{<name>, <value>}".
+ * @param data The data structure to update.
+ * @return 0 if the element is valid and updated successfully, 1 otherwise.
+ */
 int	check_element(char **val, t_data *data)
 {
 	static char	*elem[] = {"NO", "SO", "WE", "EA", "F", "C", NULL};
@@ -60,7 +127,15 @@ int	check_element(char **val, t_data *data)
 	return (1);
 }
 
-int	get_elements(char **map, t_data *data)
+/**
+ * @brief Parses the elements in the map and stores them in the data structure.
+
+ * @param map The map to parse.
+ * @param data The data structure to store the elements.
+ * @return the index of the first non-empty line if the map is found, -1 
+ * otherwise.
+ */
+int	parse_elements(char **map, t_data *data)
 {
 	char		**val;
 	int			i;
@@ -86,20 +161,4 @@ int	get_elements(char **map, t_data *data)
 	if (values_stored != 6)
 		return (error("missing map element", NULL), -1);
 	return (error("no map found", NULL), -1);
-}
-
-int	init_data(t_data *data)
-{
-	char	**map;
-	int		map_start;
-
-	map = get_file(data->mapfd);
-	if (!map)
-		return (error("init_data()", "error retrieving file"));
-	close(data->mapfd);
-	map_start = get_elements(map, data);
-	if (map_start == -1)
-		return (freetab(map, TRUE), 1);
-	freetab(map, TRUE);
-	return (0);
 }
